Move nested definitions into `types` package namespace
