+++ src/main/kotlin/com/workos/organizations/types/OrganizationDomainDataState.kt
@@ -1,0 +1,17 @@
+package com.workos.organizations.types
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+/**
+  * The verification state of the domain
+  *
+  * @property PENDING The domain is pending verification.
+  * @property VERIFIED The domain has been verified. Pass this if you've already verified the domain through other methods.
+  */
+enum class OrganizationDomainDataState {
+  @JsonProperty("verified")
+  VERIFIED,
+
+  @JsonProperty("pending")
+  PENDING
+}
